+++ src/test/java/api/support/matchers/ActualCostRecordMatchers.java
@@ -1,7 +1,6 @@
 package api.support.matchers;
 
 import java.util.HashMap;
-import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
@@ -10,8 +9,6 @@ import java.util.stream.IntStream;
 import org.folio.circulation.domain.ItemLossType;
 import org.hamcrest.Matcher;
 import org.hamcrest.core.Is;
-
-import com.jayway.jsonpath.matchers.JsonPathMatchers;
 
 import api.support.http.IndividualResource;
 import api.support.http.ItemResource;
@@ -30,7 +27,9 @@ public class ActualCostRecordMatchers {
 
   public static Matcher<JsonObject> isActualCostRecord(IndividualResource loan, ItemResource item,
     UserResource user, ItemLossType itemLossType, String permanentLocationName,
-    IndividualResource feeFineOwner, IndividualResource feeFine) {
+    IndividualResource feeFineOwner, IndividualResource feeFine, String identifierType,
+    IndividualResource patronGroup) {
+
     JsonObject instanceJson = item.getInstance().getJson();
     JsonObject itemJson = item.getJson();
     JsonArray identifiers = instanceJson.getJsonArray("identifiers");
@@ -38,12 +37,14 @@ public class ActualCostRecordMatchers {
     List<Matcher<? super String>> identifierMatchers = IntStream.range(0, identifiers.size())
       .mapToObj(i -> {
         Map<String, Matcher<String>> matchers = new HashMap<>();
-        String currentIdentifierString = String.format("identifiers[%d]", i);
+        String currentIdentifierString = String.format("instance.identifiers[%d]", i);
         JsonObject currentIdentifierObject = (JsonObject) identifiers.getValue(i);
         matchers.put(currentIdentifierString + ".identifierTypeId",
           Is.is(currentIdentifierObject.getString("identifierTypeId")));
         matchers.put(currentIdentifierString + ".value",
           Is.is(currentIdentifierObject.getString("value")));
+        matchers.put(currentIdentifierString + ".identifierType",
+          Is.is(identifierType));
 
         return toStringMatcher(matchers);
       })
@@ -51,27 +52,36 @@ public class ActualCostRecordMatchers {
 
     JsonObject effectiveCallNumberComponents = itemJson.getJsonObject(
       "effectiveCallNumberComponents");
-    return allOf(hasJsonPath("userId", is(user.getId())),
-      hasJsonPath("userBarcode", user.getBarcode()),
-      hasJsonPath("loanId", is(loan.getId())),
-      hasJsonPath("itemLossType", itemLossType.getValue()),
-      hasJsonPath("dateOfLoss", loan.getJson().getJsonObject("agedToLostDelayedBilling")
+    return allOf(hasJsonPath("user.id", is(user.getId())),
+      hasJsonPath("user.barcode", user.getBarcode()),
+      hasJsonPath("user.patronGroupId", user.getJson().getString("patronGroup")),
+      hasJsonPath("user.patronGroup", patronGroup.getJson().getString("group")),
+      hasJsonPath("loan.id", is(loan.getId())),
+      hasJsonPath("lossType", itemLossType.getValue()),
+      hasJsonPath("lossDate", loan.getJson().getJsonObject("agedToLostDelayedBilling")
         .getString("agedToLostDate")),
-      hasJsonPath("title", instanceJson.getString("title")),
-      hasJsonPath("itemBarcode", item.getBarcode()),
-      hasJsonPath("loanType", "Can Circulate"),
-      hasJsonPath("effectiveCallNumberComponents.callNumber",
+      hasJsonPath("instance.title", instanceJson.getString("title")),
+      hasJsonPath("item.barcode", item.getBarcode()),
+      hasJsonPath("item.materialTypeId", itemJson.getString("materialTypeId")),
+      hasJsonPath("item.materialType", "Book"),
+      hasJsonPath("item.loanTypeId", itemJson.getString("permanentLoanTypeId")),
+      hasJsonPath("item.loanType", "Can Circulate"),
+      hasJsonPath("item.effectiveCallNumberComponents.callNumber",
         effectiveCallNumberComponents.getString("callNumber")),
       allOfPaths(identifierMatchers),
-      hasJsonPath("effectiveCallNumberComponents.prefix",
+      hasJsonPath("item.effectiveCallNumberComponents.prefix",
         effectiveCallNumberComponents.getString("prefix")),
-      hasJsonPath("effectiveCallNumberComponents.suffix",
+      hasJsonPath("item.effectiveCallNumberComponents.suffix",
         effectiveCallNumberComponents.getString("suffix")),
-      hasJsonPath("permanentItemLocation", permanentLocationName),
-      hasJsonPath("feeFineOwnerId", is(feeFineOwner.getId())),
-      hasJsonPath("feeFineOwner", feeFineOwner.getJson().getString("owner")),
-      hasJsonPath("feeFineTypeId", is(feeFine.getId())),
-      hasJsonPath("feeFineType", feeFine.getJson().getString("feeFineType")));
+      hasJsonPath("item.permanentLocation", permanentLocationName),
+      hasJsonPath("item.volume", itemJson.getString("volume")),
+      hasJsonPath("item.chronology", itemJson.getString("chronology")),
+      hasJsonPath("item.enumeration", itemJson.getString("enumeration")),
+      hasJsonPath("item.copyNumber", itemJson.getString("copyNumber")),
+      hasJsonPath("feeFine.ownerId", is(feeFineOwner.getId())),
+      hasJsonPath("feeFine.owner", feeFineOwner.getJson().getString("owner")),
+      hasJsonPath("feeFine.typeId", is(feeFine.getId())),
+      hasJsonPath("feeFine.type", feeFine.getJson().getString("feeFineType")));
   }
 
 }