+++ src/test/java/org/folio/circulation/resources/RequestByInstanceIdResourceTests.java
@@ -52,11 +52,11 @@ public class RequestByInstanceIdResourceTests {
     for (JsonObject itemRequestJson: requestRepresentations) {
       assertEquals(item.getItemId(), itemRequestJson.getString("itemId"));
       if (i == 0)
-        assertEquals(RequestType.RECALL.getValue(), itemRequestJson.getString("requestType"));
-      if (i == 1)
-        assertEquals(RequestType.HOLD.getValue(), itemRequestJson.getString("requestType"));
-      if (i == 2)
         assertEquals(RequestType.PAGE.getValue(), itemRequestJson.getString("requestType"));
+      if (i == 1)
+        assertEquals(RequestType.RECALL.getValue(), itemRequestJson.getString("requestType"));
+      if (i == 2)
+        assertEquals(RequestType.HOLD.getValue(), itemRequestJson.getString("requestType"));
       i++;
 
       if (i > 2) {