+++ NEWS.md
@@ -1,5 +1,6 @@
 ## 23.2.0 IN PROGRESS
 
+* Supports users interface 14.2, 15.0, 16.0 (CIRC-1572)
 * Upgrade `calendar` interface to `5.0`
 
 ## 23.1.0 2022-06-28